refactor: unify footer component across pages
