+++ resources/js/main/components/Footer.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Icon from '../../../components/AppIcon';
+import Icon from './AppIcon';
 
 const Footer = () => {
   const currentYear = new Date()?.getFullYear();
